// src/generated/cs_gameevents.d.ts
export enum NET_Messages {
    net_NOP = 0,
    net_Disconnect = 1,
    net_SplitScreenUser = 3,
    net_Tick = 4,
    net_StringCmd = 5,
    net_SetConVar = 6,
    net_SignonState = 7,
    net_SpawnGroup_Load = 8,
    net_SpawnGroup_ManifestUpdate = 9,
    net_SpawnGroup_SetCreationTick = 11,
    net_SpawnGroup_Unload = 12,
    net_SpawnGroup_LoadCompleted = 13,
    net_DebugOverlay = 15
}
// ...
export class CMsgPlayerInfo implements ICMsgPlayerInfo {

    /**
     * Constructs a new CMsgPlayerInfo.
     * @param [properties] Properties to set
     */
    constructor(properties?: ICMsgPlayerInfo);

    /** CMsgPlayerInfo name. */
    public name: string;

    /** CMsgPlayerInfo xuid. */
    public xuid: (number|Long);

    /** CMsgPlayerInfo userid. */
    public userid: number;

    /** CMsgPlayerInfo steamid. */
    public steamid: (number|Long);

    /** CMsgPlayerInfo fakeplayer. */
    public fakeplayer: boolean;

    /** CMsgPlayerInfo ishltv. */
    public ishltv: boolean;

    /**
     * Creates a new CMsgPlayerInfo instance using the specified properties.
     * @param [properties] Properties to set
     * @returns CMsgPlayerInfo instance
     */
    public static create(properties?: ICMsgPlayerInfo): CMsgPlayerInfo;

    /**
     * Encodes the specified CMsgPlayerInfo message. Does not implicitly {@link CMsgPlayerInfo.verify|verify} messages.
     * @param message CMsgPlayerInfo message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: ICMsgPlayerInfo, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Encodes the specified CMsgPlayerInfo message, length delimited. Does not implicitly {@link CMsgPlayerInfo.verify|verify} messages.
     * @param message CMsgPlayerInfo message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encodeDelimited(message: ICMsgPlayerInfo, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a CMsgPlayerInfo message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns CMsgPlayerInfo
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): CMsgPlayerInfo;

    /**
     * Decodes a CMsgPlayerInfo message from the specified reader or buffer, length delimited.
     * @param reader Reader or buffer to decode from
     * @returns CMsgPlayerInfo
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): CMsgPlayerInfo;

    /**
     * Verifies a CMsgPlayerInfo message.
     * @param message Plain object to verify
     * @returns `null` if valid, otherwise the reason why it is not
     */
    public static verify(message: { [k: string]: any }): (string|null);

    /**
     * Creates a CMsgPlayerInfo message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns CMsgPlayerInfo
     */
    public static fromObject(object: { [k: string]: any }): CMsgPlayerInfo;

    /**
     * Creates a plain object from a CMsgPlayerInfo message. Also converts values to other types if specified.
     * @param message CMsgPlayerInfo
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: CMsgPlayerInfo, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this CMsgPlayerInfo to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };

    /**
     * Gets the default type url for CMsgPlayerInfo
     * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
     * @returns The default type url
     */
    public static getTypeUrl(typeUrlPrefix?: string): string;
}
// ...
export class CNETMsg_NOP implements ICNETMsg_NOP {

    /**
     * Constructs a new CNETMsg_NOP.
     * @param [properties] Properties to set
     */
    constructor(properties?: ICNETMsg_NOP);

    /**
     * Creates a new CNETMsg_NOP instance using the specified properties.
     * @param [properties] Properties to set
     * @returns CNETMsg_NOP instance
     */
    public static create(properties?: ICNETMsg_NOP): CNETMsg_NOP;

    /**
     * Encodes the specified CNETMsg_NOP message. Does not implicitly {@link CNETMsg_NOP.verify|verify} messages.
     * @param message CNETMsg_NOP message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: ICNETMsg_NOP, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Encodes the specified CNETMsg_NOP message, length delimited. Does not implicitly {@link CNETMsg_NOP.verify|verify} messages.
     * @param message CNETMsg_NOP message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encodeDelimited(message: ICNETMsg_NOP, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a CNETMsg_NOP message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns CNETMsg_NOP
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): CNETMsg_NOP;

    /**
     * Decodes a CNETMsg_NOP message from the specified reader or buffer, length delimited.
     * @param reader Reader or buffer to decode from
     * @returns CNETMsg_NOP
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): CNETMsg_NOP;

    /**
     * Verifies a CNETMsg_NOP message.
     * @param message Plain object to verify
     * @returns `null` if valid, otherwise the reason why it is not
     */
    public static verify(message: { [k: string]: any }): (string|null);

    /**
     * Creates a CNETMsg_NOP message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns CNETMsg_NOP
     */
    public static fromObject(object: { [k: string]: any }): CNETMsg_NOP;

    /**
     * Creates a plain object from a CNETMsg_NOP message. Also converts values to other types if specified.
     * @param message CNETMsg_NOP
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: CNETMsg_NOP, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this CNETMsg_NOP to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };

    /**
     * Gets the default type url for CNETMsg_NOP
     * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
     * @returns The default type url
     */
    public static getTypeUrl(typeUrlPrefix?: string): string;
}
// ...
export class CNETMsg_Disconnect implements ICNETMsg_Disconnect {

    /**
     * Constructs a new CNETMsg_Disconnect.
     * @param [properties] Properties to set
     */
    constructor(properties?: ICNETMsg_Disconnect);

    /** CNETMsg_Disconnect reason. */
    public reason: ENetworkDisconnectionReason;

    /**
     * Creates a new CNETMsg_Disconnect instance using the specified properties.
     * @param [properties] Properties to set
     * @returns CNETMsg_Disconnect instance
     */
    public static create(properties?: ICNETMsg_Disconnect): CNETMsg_Disconnect;

    /**
     * Encodes the specified CNETMsg_Disconnect message. Does not implicitly {@link CNETMsg_Disconnect.verify|verify} messages.
     * @param message CNETMsg_Disconnect message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: ICNETMsg_Disconnect, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Encodes the specified CNETMsg_Disconnect message, length delimited. Does not implicitly {@link CNETMsg_Disconnect.verify|verify} messages.
     * @param message CNETMsg_Disconnect message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encodeDelimited(message: ICNETMsg_Disconnect, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a CNETMsg_Disconnect message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns CNETMsg_Disconnect
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): CNETMsg_Disconnect;

    /**
     * Decodes a CNETMsg_Disconnect message from the specified reader or buffer, length delimited.
     * @param reader Reader or buffer to decode from
     * @returns CNETMsg_Disconnect
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): CNETMsg_Disconnect;

    /**
     * Verifies a CNETMsg_Disconnect message.
     * @param message Plain object to verify
     * @returns `null` if valid, otherwise the reason why it is not
     */
    public static verify(message: { [k: string]: any }): (string|null);

    /**
     * Creates a CNETMsg_Disconnect message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns CNETMsg_Disconnect
     */
    public static fromObject(object: { [k: string]: any }): CNETMsg_Disconnect;

    /**
     * Creates a plain object from a CNETMsg_Disconnect message. Also converts values to other types if specified.
     * @param message CNETMsg_Disconnect
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: CNETMsg_Disconnect, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this CNETMsg_Disconnect to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };

    /**
     * Gets the default type url for CNETMsg_Disconnect
     * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
     * @returns The default type url
     */
    public static getTypeUrl(typeUrlPrefix?: string): string;
}