// src/generated/netmessages.d.ts
export enum SVC_Messages {
    svc_ServerInfo = 40,
    svc_FlattenedSerializer = 41,
    svc_ClassInfo = 42,
    svc_SetPause = 43,
    svc_CreateStringTable = 44,
    svc_UpdateStringTable = 45,
    svc_VoiceInit = 46,
    svc_VoiceData = 47,
    svc_Print = 48,
    svc_Sounds = 49,
    svc_SetView = 50,
    svc_ClearAllStringTables = 51,
    svc_CmdKeyValues = 52,
    svc_BSPDecal = 53,
    svc_SplitScreen = 54,
    svc_PacketEntities = 55,
    svc_Prefetch = 56,
    svc_Menu = 57,
    svc_GetCvarValue = 58,
    svc_StopSound = 59,
    svc_PeerList = 60,
    svc_PacketReliable = 61,
    svc_HLTVStatus = 62,
    svc_ServerSteamID = 63,
    svc_FullFrameSplit = 70,
    svc_RconServerDetails = 71,
    svc_UserMessage = 72,
    svc_HltvReplay = 73,
    svc_Broadcast_Command = 74,
    svc_HltvFixupOperatorStatus = 75
}
// ...
export class CSVCMsg_ServerInfo implements ICSVCMsg_ServerInfo {

    /**
     * Constructs a new CSVCMsg_ServerInfo.
     * @param [properties] Properties to set
     */
    constructor(properties?: ICSVCMsg_ServerInfo);

    /** CSVCMsg_ServerInfo protocol. */
    public protocol: number;

    /** CSVCMsg_ServerInfo serverCount. */
    public serverCount: number;

    /** CSVCMsg_ServerInfo isDedicated. */
    public isDedicated: boolean;

    /** CSVCMsg_ServerInfo isHltv. */
    public isHltv: boolean;

    /** CSVCMsg_ServerInfo cOs. */
    public cOs: number;

    /** CSVCMsg_ServerInfo maxClients. */
    public maxClients: number;

    /** CSVCMsg_ServerInfo maxClasses. */
    public maxClasses: number;

    /** CSVCMsg_ServerInfo playerSlot. */
    public playerSlot: number;

    /** CSVCMsg_ServerInfo tickInterval. */
    public tickInterval: number;

    /** CSVCMsg_ServerInfo gameDir. */
    public gameDir: string;

    /** CSVCMsg_ServerInfo mapName. */
    public mapName: string;

    /** CSVCMsg_ServerInfo skyName. */
    public skyName: string;

    /** CSVCMsg_ServerInfo hostName. */
    public hostName: string;

    /** CSVCMsg_ServerInfo addonName. */
    public addonName: string;

    /** CSVCMsg_ServerInfo gameSessionConfig. */
    public gameSessionConfig?: (ICSVCMsg_GameSessionConfiguration|null);

    /** CSVCMsg_ServerInfo gameSessionManifest. */
    public gameSessionManifest: Uint8Array;

    /**
     * Creates a new CSVCMsg_ServerInfo instance using the specified properties.
     * @param [properties] Properties to set
     * @returns CSVCMsg_ServerInfo instance
     */
    public static create(properties?: ICSVCMsg_ServerInfo): CSVCMsg_ServerInfo;

    /**
     * Encodes the specified CSVCMsg_ServerInfo message. Does not implicitly {@link CSVCMsg_ServerInfo.verify|verify} messages.
     * @param message CSVCMsg_ServerInfo message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: ICSVCMsg_ServerInfo, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Encodes the specified CSVCMsg_ServerInfo message, length delimited. Does not implicitly {@link CSVCMsg_ServerInfo.verify|verify} messages.
     * @param message CSVCMsg_ServerInfo message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encodeDelimited(message: ICSVCMsg_ServerInfo, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a CSVCMsg_ServerInfo message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns CSVCMsg_ServerInfo
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): CSVCMsg_ServerInfo;

    /**
     * Decodes a CSVCMsg_ServerInfo message from the specified reader or buffer, length delimited.
     * @param reader Reader or buffer to decode from
     * @returns CSVCMsg_ServerInfo
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): CSVCMsg_ServerInfo;

    /**
     * Verifies a CSVCMsg_ServerInfo message.
     * @param message Plain object to verify
     * @returns `null` if valid, otherwise the reason why it is not
     */
    public static verify(message: { [k: string]: any }): (string|null);

    /**
     * Creates a CSVCMsg_ServerInfo message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns CSVCMsg_ServerInfo
     */
    public static fromObject(object: { [k: string]: any }): CSVCMsg_ServerInfo;

    /**
     * Creates a plain object from a CSVCMsg_ServerInfo message. Also converts values to other types if specified.
     * @param message CSVCMsg_ServerInfo
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: CSVCMsg_ServerInfo, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this CSVCMsg_ServerInfo to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };

    /**
     * Gets the default type url for CSVCMsg_ServerInfo
     * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
     * @returns The default type url
     */
    public static getTypeUrl(typeUrlPrefix?: string): string;
}